fix(intel): address code-review-expert findings
P1 — prune({limit: 0}) rescheduled itself forever: take(0) returns [],
and the unguarded `stale.length >= batch` read 0 >= 0 as a full batch,
so the mutation looped in production deleting nothing. Verified by
execution before fixing ({deleted:0, rescheduled:true}); batch now
floors to 1 and a regression test pins it.

P2 — get-intel-timeline had no rate policy on the reasoning that it
spends no provider money. True, but it spends the resource that scales
with retention: Convex reads whole documents, so one limit=200 call
scans 800 rows each carrying a 512-float embedding the projection
discards. Now 120/min.

P2 — split intel-history-client.ts: embedding (paid upstream, own
budget and failure mode, and the half that grows when the model is
versioned in #5742) moved to intel-history-embed.ts; the client is now
Convex transport only.

P3 — dropped the SeedHistoryError and toIntelHistoryRecord exports
(nothing imported either); scopeBody now forwards a negative from/to
verbatim instead of silently discarding it, which is what the MCP layer
already promised callers.

Part of #5694.
